Fix code in modules/leave and change logic file
